Ensure transitive legacy support libraries are included in `jetify_include_list`.
It seems specifying top level artifacts to be jetified in `jetify_include_list` is not sufficient as `rules_jvm_external` does not properly consider transitive dependencies and ensures they are jetified. To overcome this, we map the newer jetifier versions to old ones and ensure they are included in `jetify_include_list`.

Additionally `ResolvedComponentVisitor` is decoupled form `ResolvedDependencyResult` by moving dependency notation encoding to callbak.

Signed-off-by: arunkumar9t2 <arunk.beece@gmail.com>